add clean; use post_command; add --check; etc

+++ scripts/luaxake/luaxake-compile.lua
@@ -2,8 +2,10 @@ local M = {}
 local lfs = require "lfs"
 local error_logparser = require("make4ht-errorlogparser")
 local pl = require "penlight"
-local mkutils = require("mkutils")
 local path = pl.path
+local pldir = pl.dir
+local pltable = require "pl.tablex"
+local plfile = pl.file
 local html = require "luaxake-transform-html"
 local files = require "luaxake-files"      -- for get_metadaa
 local socket = require "socket"
@@ -45,41 +47,83 @@ local function copy_table(tbl)
   return t
 end
 
+-- Function to find the first table with a given key/value using Penlight
+local function find_entry(array, key, value)
+  for _, entry in ipairs(array) do
+      if entry[key] == value then
+          return entry  -- Return the first matching entry
+      end
+  end
+  return nil  -- Return nil if no match is found
+end
+
+-- HACK: these need to be global; TODO: fix!
+function process_html(file)
+  return html.process(file)
+end
+
+function move_to_downloads(file, cmd_meta, root_dir)
+  local folder = string.format("%s/%s/%s",root_dir, cmd_meta.download_folder, file.dir)
+  
+  require 'pl.pretty'.dump(cmd_meta)
+  require 'pl.pretty'.dump(file)
+  local src = find_entry(file.output_files, "extensionlong", cmd_meta.extension)
+  -- local src = file.output_files[1].absolute_path    -- TODO: fix
+  
+  local tgt = string.format("%s/%s.%s", folder, file.basename, src.extension)
+  -- require 'pl.pretty'.dump(src)
+  if src and path.exists(src.absolute_path) then
+    log:infof("Moving %s to %s", src.absolute_path, tgt)
+    pldir.makepath(folder)
+    plfile.copy(src.absolute_path, tgt)
+  else
+    log:warningf("No output file found for ",file.relative_path)
+  end
+  return 1, 'NOK'
+end
+
+
 --- run a command
 --- @param file metadata file on which the command should be run
 --- @param compilers [compiler] list of compilers
 --- @param compile_sequence table sequence of keys from the compilers table to be executed
 --- @return [compile_info] statuses information from the commands
-local function compile(file, compilers, compile_sequence)
+local function compile(file, compilers, compile_sequence, only_check)
+  only_check = only_check or false
   
   --
   -- WARNING: (tex-)compilation STARTS IN THE SUBFOLDER !!!
   --
   local current_dir = lfs.currentdir()
+  log:tracef("Changing directory to %s (for actual compilations, from %s)",file.absolute_dir,current_dir)
   lfs.chdir(file.absolute_dir)
 
   local statuses = {}
   local FAIL = false
 
   -- Start ALL compilations for this file, in the correct order; stop as soon as one fails...
+  -- NOTE: extension is a bad name, it's rather  'compiler'
   for _, extension in ipairs(compile_sequence) do
     local command_metadata = compilers[extension]
 
     if not command_metadata then
-      log:errorf("No compiler defined for %s (%s)",extension,file.relative_path)
-      error("No compiler defined for "..extension)
+      log:errorf("No compiler defined for %s (%s); SKIPPING",extension,file.relative_path)
+      goto endofthiscompilation  -- nice: a goto-statement !!!
     end
 
     local output_file = file.filename:gsub("tex$", extension)
-    if command_metadata and command_metadata.check_file then
-      -- sometimes compiler wants to check for the output file (like for sagetex.sage),
-      if not mkutils.file_exists(output_file) then
-        -- ignore this command if the file doesn't exist
-        command_metadata = nil
-      end
+
+    -- sometimes compiler wants to check for the output file (like for sagetex.sage),
+    if command_metadata.check_file and not path.exists(output_file) then
+      log:debugf("Skipping compilation because file %s does not exist",output_file)
+      goto endofthiscompilation  -- nice: a goto-statement !!!
     end
-    -- if command_metadata and output.needs_compilation then
-    if command_metadata then
+    
+    -- if not output_file.needs_compilation then
+    --   log:debugf("Skipping compilation file %s is uptodate",output_file)
+    --   goto endofthiscompilation  -- nice: a goto-statement !!!
+    -- end
+
       local command_template = command_metadata.command
       -- we need to make a copy of file metadata to insert some additional fields without modification of the original
       -- log:debug("Command " .. command_template)
@@ -89,23 +133,31 @@ local function compile(file, compilers, compile_sequence)
       tpl_table.make4ht_mode = config.make4ht_mode
       local command = prepare_command(tpl_table, command_template)
       local start_time =  socket.gettime()
+      local compilation_time = 0
+      local status = 0
+      local output = ""
 
-
-      log:info("Starting " .. command )
+      if only_check then
+        log:info("Running in check-modus: SKIPPING " .. command )
+      else
       -- we reuse this file from make4ht's mkutils.lua
       local f = io.popen(command, "r")
-      local output = f:read("*all")
+      output = f:read("*all")
       -- rc will contain return codes of the executed command
       local rc =  {f:close()}
       -- the status code is on the third position 
       -- https://stackoverflow.com/a/14031974/2467963
-      local status = rc[3]
+      status = rc[3]
       local end_time = socket.gettime()
-      local compilation_time = end_time - start_time
+      compilation_time = end_time - start_time
+
       if status ~= command_metadata.status then
-        log:error("Compilation failed: returns " .. (status or "") ..", but expected ".. command_metadata.status)
-        FAIL = true   -- continue for now, to collect logging etc. ...
+        -- error will be handled and logged further down!
+        log:errorf("Compilation failed: returns %d (and not %d) after %3f seconds", status, command_metadata.status,compilation_time)
       end
+      end
+
+
       --- @class compile_info
       --- @field output_file string output file name
       --- @field command string executed command
@@ -124,31 +176,57 @@ local function compile(file, compilers, compile_sequence)
         info.errors = test_log_file(file.basename .. ".log")
       end
 
-      -- store outputfiles with metadata; TODO: check/fix absolute_path
-      local ofile = files.get_metadata(file.absolute_dir, output_file)
+      -- store outputfiles with metadata
+      local ofile = files.get_metadata(output_file)
 
       log:debug("Adding outputfile "..ofile.relative_path.. " to "..file.relative_path)
       -- require 'pl.pretty'.dump(ofile)
       table.insert(file.output_files,ofile) 
       -- require 'pl.pretty'.dump(file)
 
-      if command_metadata.process_html then
-        info.html_processing_status, info.html_processing_message = html.process(file)
-        if not info.html_processing_status then
-          log:error("Error in HTML post processing: " .. (info.html_processing_message or ""))
+      if command_metadata.post_command then
+        local cmd = command_metadata.post_command
+        log:infof("Postprocessing: %s", cmd)
+
+        info.post_status, info.post_message = _G[cmd](file, command_metadata, current_dir)     -- lua way of calling the function whose name is in 'cmd'
+        if not info.post_status then
+          log:error("Error in HTML post processing: " .. info.post_message)
         end
       end
+
+      -- OBSOLETE, see above 
+      -- if command_metadata.process_html then
+      --   log:info("Postprocessing html")
+
+      --   info.html_processing_status, info.html_processing_message = process_html(file)
+      --   if not info.html_processing_status then
+      --     log:error("Error in HTML post processing: " .. (info.html_processing_message or ""))
+      --   end
+      -- end
+
+
       table.insert(statuses, info)
       log:info(string.format("Compilation of %s took %.1f seconds (%.20s)", output_file, compilation_time, file.title))
 
-      if FAIL then
+      if status == command_metadata.fatal_status then
+        log:warning("Skipping further compilations for %s after error",relative_file)
         break   -- STOP FURTHER COMPILATION
       end
+    --end
+    log:tracef("Ended compilation %s", extension)
+    ::endofthiscompilation::
   end
-  end
+  ::endofallcompilations::
   lfs.chdir(current_dir)
+
+  -- if dump_metadata then
+  --   log:debug("Dumping new metadata for ".. relative_file )
+  --   require 'pl.pretty'.dump(metadata)
+  -- end
+
   return statuses
 end
+
 
 --- print error messages parsed from the LaTeX log
 ---@param errors table
@@ -167,16 +245,21 @@ end
 
 --- remove temporary files
 ---@param basefile metadata 
----@param extensions table list of extensions of files to be removed
-local function clean(basefile, extensions)
+---@param extensions table    list of extensions of files to be removed
+---@return  number nfiles     number of files removed
+local function clean(basefile, extensions, only_check)
+  only_check = only_check or false
+  local nfiles = 0
   local basename = path.splitext(basefile.absolute_path)
+  log:tracef("%s temp files for %s (%s)", (only_check and "Would remove" or "Removing"), basename, basefile.absolute_path)
   for _, ext in ipairs(extensions) do
     local filename = basename .. "." .. ext
-    if mkutils.file_exists(filename) then
-      log:debug("Removing temp file: " .. filename)
-      os.remove(filename)
+    if path.exists(filename) then
+      log:debugf("%s  %s file %s", (only_check and "Would remove" or "Removing") ,ext, filename)
+      if not only_check then os.remove(filename); nfiles = nfiles + 1 end
     end
   end
+  return nfiles
 end
 
 M.compile      = compile
